Fix CounterController Metric value type & add tests
 - The metric value type used in the CounterController is set to use `Long` instead of `Double`
 - Add CounterControllerTests

This resolves #526 and 460

+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/controller/CounterController.java
@@ -46,6 +46,7 @@ import org.springframework.web.bind.annotation.RestController;
  *
  * @author Eric Bottard
  * @author Mark Fisher
+ * @author Ilayaperumal Gopinathan
  */
 @RestController
 @RequestMapping("/metrics/counters")
@@ -56,10 +57,10 @@ public class CounterController {
 
 	private final MetricRepository metricRepository;
 
-	private final ResourceAssembler<Metric<Double>, CounterResource> counterResourceAssembler =
+	private final ResourceAssembler<Metric<Long>, CounterResource> counterResourceAssembler =
 			new DeepCounterResourceAssembler();
 
-	protected final ResourceAssembler<Metric<Double>, ? extends MetricResource> shallowResourceAssembler =
+	protected final ResourceAssembler<Metric<Long>, ? extends MetricResource> shallowResourceAssembler =
 			new ShallowMetricResourceAssembler();
 
 	/**
@@ -78,12 +79,12 @@ public class CounterController {
 	@RequestMapping(value = "", method = RequestMethod.GET)
 	public PagedResources<? extends MetricResource> list(
 			Pageable pageable,
-			PagedResourcesAssembler<Metric<Double>> pagedAssembler,
+			PagedResourcesAssembler<Metric<Long>> pagedAssembler,
 			@RequestParam(value = "detailed", defaultValue = "false") boolean detailed) {
 		/* Page */ Iterable<Metric<?>> metrics = metricRepository.findAll(/* pageable */);
-		List<Metric<Double>> content = filterCounters(metrics);
-		Page<Metric<Double>> page = new PageImpl<>(content);
-		ResourceAssembler<Metric<Double>, ? extends MetricResource> assemblerToUse =
+		List<Metric<Long>> content = filterCounters(metrics);
+		Page<Metric<Long>> page = new PageImpl<>(content);
+		ResourceAssembler<Metric<Long>, ? extends MetricResource> assemblerToUse =
 				detailed ? counterResourceAssembler : shallowResourceAssembler;
 		return pagedAssembler.toResource(page, assemblerToUse);
 	}
@@ -93,7 +94,7 @@ public class CounterController {
 	 */
 	@RequestMapping(value = "/{name}", method = RequestMethod.GET)
 	public CounterResource display(@PathVariable("name") String name) {
-		Metric<Double> c = findCounter(name);
+		Metric<Long> c = findCounter(name);
 		return counterResourceAssembler.toResource(c);
 	}
 
@@ -103,7 +104,7 @@ public class CounterController {
 	@RequestMapping(value = "/{name}", method = RequestMethod.DELETE)
 	@ResponseStatus(HttpStatus.OK)
 	protected void delete(@PathVariable("name") String name) {
-		Metric<Double> c = findCounter(name);
+		Metric<Long> c = findCounter(name);
 		metricRepository.reset(c.getName());
 	}
 
@@ -111,9 +112,9 @@ public class CounterController {
 	 * Find a given counter, taking care of name conversion between the Spring Boot domain and our domain.
 	 * @throws MetricsMvcEndpoint.NoSuchMetricException if the counter does not exist
 	 */
-	private Metric<Double> findCounter(@PathVariable("name") String name) {
+	private Metric<Long> findCounter(@PathVariable("name") String name) {
 		@SuppressWarnings("unchecked")
-		Metric<Double> c = (Metric<Double>) metricRepository.findOne(COUNTER_PREFIX + name);
+		Metric<Long> c = (Metric<Long>) metricRepository.findOne(COUNTER_PREFIX + name);
 		if (c == null) {
 			throw new MetricsMvcEndpoint.NoSuchMetricException(name);
 		}
@@ -142,19 +143,19 @@ public class CounterController {
 	 * @author Eric Bottard
 	 */
 	static class ShallowMetricResourceAssembler extends
-			ResourceAssemblerSupport<Metric<Double>, MetricResource> {
+			ResourceAssemblerSupport<Metric<Long>, MetricResource> {
 
 		public ShallowMetricResourceAssembler() {
 			super(CounterController.class, MetricResource.class);
 		}
 
 		@Override
-		public MetricResource toResource(Metric<Double> entity) {
+		public MetricResource toResource(Metric<Long> entity) {
 			return createResourceWithId(entity.getName().substring(COUNTER_PREFIX.length()), entity);
 		}
 
 		@Override
-		protected MetricResource instantiateResource(Metric<Double> entity) {
+		protected MetricResource instantiateResource(Metric<Long> entity) {
 			return new MetricResource(entity.getName().substring(COUNTER_PREFIX.length()));
 		}
 
@@ -166,19 +167,19 @@ public class CounterController {
 	 * @author Eric Bottard
 	 */
 	static class DeepCounterResourceAssembler extends
-			ResourceAssemblerSupport<Metric<Double>, CounterResource> {
+			ResourceAssemblerSupport<Metric<Long>, CounterResource> {
 
 		public DeepCounterResourceAssembler() {
 			super(CounterController.class, CounterResource.class);
 		}
 
 		@Override
-		public CounterResource toResource(Metric<Double> entity) {
+		public CounterResource toResource(Metric<Long> entity) {
 			return createResourceWithId(entity.getName().substring(COUNTER_PREFIX.length()), entity);
 		}
 
 		@Override
-		protected CounterResource instantiateResource(Metric<Double> entity) {
+		protected CounterResource instantiateResource(Metric<Long> entity) {
 			return new CounterResource(entity.getName().substring(COUNTER_PREFIX.length()), entity.getValue().longValue());
 		}
